Meter: apply zero-energy guard at meter call site, not shared helper
