+++ meter/measurement/energy.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/evcc-io/evcc/api"
 	"github.com/evcc-io/evcc/plugin"
 )
 
@@ -35,21 +34,5 @@ func (cc *Energy) Configure(ctx context.Context) (
 		return nil, nil, nil, fmt.Errorf("returnEnergy: %w", err)
 	}
 
-	return powerG, nonZeroEnergy(energyG), nonZeroEnergy(returnG), nil
-}
-
-// nonZeroEnergy reports a zero lifetime energy reading as api.ErrNotAvailable.
-// Some inverters reset their total counter to 0 at night; this keeps the last value (#30951).
-func nonZeroEnergy(g func() (float64, error)) func() (float64, error) {
-	if g == nil {
-		return nil
-	}
-
-	return func() (float64, error) {
-		f, err := g()
-		if err == nil && f == 0 {
-			return 0, api.ErrNotAvailable
-		}
-		return f, err
-	}
+	return powerG, energyG, returnG, nil
 }
